Moved providers component to the app directory
