+++ app/(auth)/sign-in/page.tsx
@@ -5,7 +5,7 @@ import {getProviders} from "next-auth/react";
 import {getServerAuthSession} from "~/server/auth";
 import {Button} from "~/components/ui/button";
 import {Heading} from "~/components/ui/heading";
-import {SignInProviders} from "~/components/auth/sign-in-providers";
+import {SignInProviders} from "./sign-in-providers";
 
 export default async function Page() {
   const session = await getServerAuthSession();
